Increase the timing resolution for TimerResult

+++ source/Glimpse.Core/Extensibility/IExecutionTimer.cs
@@ -7,9 +7,9 @@ namespace Glimpse.Core.Extensibility
         TimerResult<T> Time<T>(Func<T> func);
         
         TimerResult Time(Action action);
-        
-        long Start();
-        
-        TimerResult Stop(long offset);
+
+        double Start();
+
+        TimerResult Stop(double offset);
     }
 }
