+++ source/Glimpse.Mvc3/AlternateImplementation/ActionInvokerStateMixin.cs
@@ -2,14 +2,14 @@
 {
     public interface IActionInvokerStateMixin
     {
-        long Offset { get; set; }
+        double Offset { get; set; }
 
         ActionInvoker.InvokeActionMethod.Arguments Arguments { get; set; }
     }
 
     public class ActionInvokerStateMixin : IActionInvokerStateMixin
     {
-        public long Offset { get; set; }
+        public double Offset { get; set; }
 
         public ActionInvoker.InvokeActionMethod.Arguments Arguments { get; set; }
     }
